experimental-service: Combine schema and tool handler
